+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     they are properly detected. [GH-683]
   - Fix issue where VM import sometimes made strange VirtualBox folder
     layouts. [GH-669]
+  - Call proper `id` command on Solaris. [GH-679]
 
 ## 0.9.3 (January 24, 2012)